Fix GRU and LSTM implementations for sample rate correction modes
GRU:
The issue was that for the case where SampleRateCorrectionMode was not None,
the values for h(t-1) in extendedHt1 weren't actually the ones in the `outs`
vector as set by `processDelay`.

LSTM:
Similar issue. When SampleRateCorrectionMode was not None, the h(t-1)
in subsequent `forward` calls did not contain the same values as the
`outs` vector as set by `processDelay`.

The `cVec` did not have the same issue because after setting the values of
`ct_delayed[delayWriteIdx]` in `computeOutputsInternal`, `processDelay` puts
the right value in `cVec`.