Eigen backend optimization for Dense, GRU and LSTM layers (#108)
* Make Dense Eigen faster

Make the Eigen implementation around 30-40% faster by fusing matrix operation.

* GRU Eigen templated implementation

Added the templated implementation for the GRU layer.
Currently, on an Apple MBA 2022, 10s 16in 16out, Release config:
- dynamic implementation is ~67x faster than realtime
- templated implementation is ~69x faster than realtime

* Fix GRU and make LSTM faster

* Add explanatory comments

* Remove auto templated parameters

* Fix comment format

* Remove unused type alias

* Fix GRU and LSTM implementations for sample rate correction modes

GRU:
The issue was that for the case where SampleRateCorrectionMode was not None,
the values for h(t-1) in extendedHt1 weren't actually the ones in the `outs`
vector as set by `processDelay`.

LSTM:
Similar issue. When SampleRateCorrectionMode was not None, the h(t-1)
in subsequent `forward` calls did not contain the same values as the
`outs` vector as set by `processDelay`.

The `cVec` did not have the same issue because after setting the values of
`ct_delayed[delayWriteIdx]` in `computeOutputsInternal`, `processDelay` puts
the right value in `cVec`.

* One more try for the Dense layer

Apparently, the += operator for NoAlias is about as fast as the just the = operator.

So I figured I'd give it a shot. On M2 it's yet a little faster.
I'll try it on Windows in a minute.

* Revert "One more try for the Dense layer"

This reverts commit 29bd1daa33fc5a6dce5b2f4f88db0abe75f1654c.