New definition for causal_crop
This definition makes it so that an apparent off-by-one error is
fixed by instead making the `start` index be the difference between
the target and the current length.

For example, given a tensor with 1000 elements, and a target of 970,
the function would instead crop with `x[..., 30:]`, rather than the
previous behaviour of `x[..., 29:999]`.

This seems to be more correct in that the previous behaviour takes
items from index 29 (inclusive) to index 999 (exclusive, so 998).

Meanwhile, the new behaviour makes it so that it's index 30 (inclusive)
until the end, which is the index 999.